Moved processRequest to Connection
