+++ include/network/Server.hpp
@@ -44,7 +44,6 @@ class Server
     void closeConnection(int clientNo);
     void recvData(size_t clientNo);
     void readBody(size_t clientNo);
-    void processRequest(Connection &c);
     void sendResponse(size_t clientNo);
 
   public:
