+++ include/responses/Response.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file Response.hpp
+ * @author Mehrin Firdousi (mehrinfirdousi@gmail.com)
+ * @brief This class describes an HTTP response for one request
+ * @date 2023-07-17
+ * 
+ * @copyright Copyright (c) 2023
+ * 
+ */
+
+#ifndef RESPONSE_HPP
+#define RESPONSE_HPP
+
+# include <iostream>
+
+class Response
+{
+    char *response;
+    
+};
+
+
+#endif
